# backend/nyaysetu-backend/src/main/java/com/nyaysetu/backend/controller/MeetingController.java
package com.nyaysetu.backend.controller;

import com.nyaysetu.backend.dto.CreateMeetingRequest;
import com.nyaysetu.backend.dto.JoinMeetingRequest;
import com.nyaysetu.backend.dto.MeetingResponse;
import com.nyaysetu.backend.service.MeetingService;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.UUID;

@Tag(name = "Meetings", description = "Create and join virtual meetings for hearings and consultations")
@RestController
@RequestMapping("/meetings")
@RequiredArgsConstructor
public class MeetingController {

    private final MeetingService meetingService;

    @PostMapping
    public ResponseEntity<MeetingResponse> create(@Valid  @RequestBody CreateMeetingRequest dto) {
        return ResponseEntity.ok(meetingService.createMeeting(dto));
    }

    @PostMapping("/join")
    public ResponseEntity<MeetingResponse> join(@Valid @RequestBody JoinMeetingRequest dto) {
        return ResponseEntity.ok(meetingService.joinMeeting(dto));
    }

    @PostMapping("/end/{id}")
    public ResponseEntity<MeetingResponse> end(@PathVariable UUID id) {
        return ResponseEntity.ok(meetingService.endMeeting(id));
    }

    @GetMapping("/case/{caseId}")
    public ResponseEntity<List<MeetingResponse>> byCase(@PathVariable UUID caseId) {
        return ResponseEntity.ok(meetingService.getMeetingsByCase(caseId));
    }
}