# docs/openapi.yaml
openapi: 3.0.0
info:
  title: NyaySetu Unified API
  version: 1.0.0
  description: REST API specification for the NyaySetu Digital Judiciary Platform. Covers Spring Boot backend services and Python microservices.
servers:
  - url: http://localhost:8080
    description: Spring Boot Backend Server
  - url: http://localhost:8001
    description: Python Services (NLP Orchestrator / LawGPT)
paths:
...
  /api/v1/hearings/case/{caseId}:
    get:
      tags:
        - Hearings
      summary: "Execute getCaseHearings"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: "The caseId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/my:
    get:
      tags:
        - Hearings
      summary: "Execute getMyHearings"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/schedule:
    post:
      tags:
        - Hearings
      summary: "Execute scheduleHearing"
      description: " (Served by Spring Boot backend)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                caseId: "case-uuid-here"
                hearingDate: "2026-07-15T10:00:00"
                hearingType: PRELIMINARY
                description: Initial hearing for plea
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}:
    get:
      tags:
        - Hearings
      summary: "Execute getHearing"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}/complete:
    put:
      tags:
        - Hearings
      summary: "Execute completeHearing"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}/join:
    post:
      tags:
        - Hearings
      summary: "Execute joinHearing"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}/leave:
    post:
      tags:
        - Hearings
      summary: "Execute leaveHearing"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}/outcome:
    post:
      tags:
        - Hearings
      summary: "Execute recordOutcome"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                outcome: "Hearing adjourned. Next hearing on 2026-08-01."
                nextHearingDate: "2026-08-01T10:00:00"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/hearings/{hearingId}/participants:
    get:
      tags:
        - Hearings
      summary: "Execute getParticipants"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                name: Advocate Sharma
                role: LAWYER
                email: "lawyer@example.com"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
    post:
      tags:
        - Hearings
      summary: "Execute addParticipant"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: hearingId
          in: path
          required: true
          schema:
            type: string
          description: "The hearingId identifier"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                name: Advocate Sharma
                role: LAWYER
                email: "lawyer@example.com"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
  /api/v1/meetings:
    post:
      tags:
        - Meetings
      summary: "Execute create"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/meetings/case/{caseId}:
    get:
      tags:
        - Meetings
      summary: "Execute byCase"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: "The caseId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/meetings/end/{id}:
    post:
      tags:
        - Meetings
      summary: "Execute end"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/meetings/join:
    post:
      tags:
        - Meetings
      summary: "Execute join"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Token from /api/v1/auth/login"